some warnings fixed
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@16541 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/kernel/discrete/delauney/Triangle.java
@@ -141,7 +141,7 @@ public class Triangle extends ArraySet<Pnt> {
 
     @Override
     public int hashCode () {
-        return (int)(idNumber^(idNumber>>>32));
+        return idNumber^(idNumber>>>32);
     }
 
     @Override
